# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        },
        "summary": "Create a new product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Locale for language (default: en)",
            "schema": {
              "example": "en",
              "type": "string",
              "enum": [
                "en",
                "vi"
              ]
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "Filter by category ID",
            "schema": {
              "example": "b4b2b07f-6825-402b-bd2c-f9aef8cfbba5",
              "type": "string"
            }
          },
          {
            "name": "collection_id",
            "required": false,
            "in": "query",
            "description": "Filter by collection ID",
            "schema": {
              "example": "c5c5d08f-7936-503c-ce3d-g0bgf9cfccb6",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "draft",
                "out_of_stock",
                "discontinued"
              ]
            }
          },
          {
            "name": "is_featured",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "enable_sale_tag",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "polo",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "schema": {
              "example": "created_at",
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "name",
                "price",
                "status"
              ]
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "schema": {
              "example": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully"
          }
        },
        "summary": "Get all products with filters",
        "tags": [
          "products"
        ]
      }
    },
    "/products/search": {
      "get": {
        "operationId": "ProductsController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          }
        },
        "summary": "Search products by keyword",
        "tags": [
          "products"
        ]
      }
    },
    "/products/slug/{slug}": {
      "get": {
        "operationId": "ProductsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get product by slug",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get product by ID",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          }
        },
        "summary": "Update product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          }
        },
        "summary": "Soft delete product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/stock": {
      "get": {
        "operationId": "ProductsController_getTotalStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock count retrieved"
          }
        },
        "summary": "Get total stock for product",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}/variants/{sku}/stock": {
      "patch": {
        "operationId": "ProductsController_updateVariantStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sku",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Variant stock updated"
          }
        },
        "summary": "Update variant stock",
        "tags": [
          "products"
        ]
      }
    },
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully"
          }
        },
        "summary": "Create a new category",
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_getCategories",
        "parameters": [
          {
            "name": "with_children_count",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched categories successfully"
          }
        },
        "summary": "Get all categories (Admin Management)",
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/tree": {
      "get": {
        "operationId": "CategoriesController_getTree",
        "parameters": [
          {
            "name": "active",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category tree retrieved successfully"
          }
        },
        "summary": "Get Category Tree (Frontend Navigation)",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/categories/slug/{slug}": {
      "get": {
        "operationId": "CategoriesController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "Get category by slug",
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "Get category by ID",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully"
          }
        },
        "summary": "Update category",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          }
        },
        "summary": "Delete category",
        "tags": [
          "categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ecom API",
    "description": "E-commerce REST API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductVariantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": {
              "en": "M - Black",
              "vi": "M - Đen"
            },
            "description": "Variant name in multiple languages"
          },
          "sku": {
            "type": "string",
            "example": "POLO-M-BLACK"
          },
          "price": {
            "type": "number",
            "example": 399000
          },
          "stock": {
            "type": "number",
            "example": 10
          },
          "barcode": {
            "type": "string",
            "example": "1234567890123"
          },
          "color_id": {
            "type": "string",
            "example": "44fd41a7-63b1-41f6-b05d-1935d392f1d4"
          },
          "size_id": {
            "type": "string",
            "example": "07bdcefc-da8a-4b29-9945-602abb4adc02"
          },
          "image_url": {
            "type": "string",
            "example": "https://res.cloudinary.com/.../red.jpg"
          }
        },
        "required": [
          "name",
          "sku",
          "price",
          "stock",
          "color_id",
          "size_id"
        ]
      },
      "DimensionsDto": {
        "type": "object",
        "properties": {
          "length": {
            "type": "number",
            "example": 28,
            "description": "Length in cm"
          },
          "width": {
            "type": "number",
            "example": 20,
            "description": "Width in cm"
          },
          "height": {
            "type": "number",
            "example": 2,
            "description": "Height in cm"
          }
        }
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": {
              "en": "Premium Polo Shirt",
              "vi": "Áo Polo Cao Cấp"
            },
            "description": "Product name in multiple languages"
          },
          "slug": {
            "type": "object",
            "example": {
              "en": "premium-polo-shirt",
              "vi": "ao-polo-cao-cap"
            },
            "description": "Product slug in multiple languages"
          },
          "description": {
            "type": "object",
            "example": {
              "en": "High quality cotton polo shirt...",
              "vi": "Áo polo cotton chất lượng cao..."
            },
            "description": "Product description in multiple languages"
          },
          "short_description": {
            "type": "object",
            "example": {
              "en": "Premium cotton polo",
              "vi": "Áo polo cotton cao cấp"
            },
            "description": "Short description in multiple languages"
          },
          "price": {
            "type": "number",
            "example": 399000
          },
          "sale_price": {
            "type": "number",
            "example": 349000
          },
          "cost_price": {
            "type": "number",
            "example": 200000
          },
          "images": {
            "example": [
              "https://example.com/image1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "stock_quantity": {
            "type": "number",
            "example": 50
          },
          "sku": {
            "type": "string",
            "example": "POLO-001"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890123"
          },
          "category_id": {
            "type": "string",
            "example": "b4b2b07f-6825-402b-bd2c-f9aef8cfbba5"
          },
          "tags": {
            "example": [
              "polo",
              "men",
              "premium"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "draft",
              "out_of_stock",
              "discontinued"
            ]
          },
          "is_featured": {
            "type": "boolean",
            "example": false
          },
          "enable_sale_tag": {
            "type": "boolean",
            "example": false
          },
          "meta_title": {
            "type": "object",
            "example": {
              "en": "Buy Premium Polo Shirt",
              "vi": "Mua Áo Polo Cao Cấp"
            },
            "description": "Meta title in multiple languages"
          },
          "meta_description": {
            "type": "object",
            "example": {
              "en": "High quality polo shirt...",
              "vi": "Áo polo chất lượng cao..."
            },
            "description": "Meta description in multiple languages"
          },
          "weight": {
            "type": "number",
            "example": 0.3,
            "description": "Weight in kg"
          },
          "dimensions": {
            "example": {
              "length": 28,
              "width": 20,
              "height": 2
            },
            "description": "Product dimensions in cm",
            "allOf": [
              {
                "$ref": "#/components/schemas/DimensionsDto"
              }
            ]
          }
        },
        "required": [
          "name",
          "slug",
          "price"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": {
              "en": "Premium Polo Shirt",
              "vi": "Áo Polo Cao Cấp"
            },
            "description": "Product name in multiple languages"
          },
          "slug": {
            "type": "object",
            "example": {
              "en": "premium-polo-shirt",
              "vi": "ao-polo-cao-cap"
            },
            "description": "Product slug in multiple languages"
          },
          "description": {
            "type": "object",
            "example": {
              "en": "High quality cotton polo shirt...",
              "vi": "Áo polo cotton chất lượng cao..."
            },
            "description": "Product description in multiple languages"
          },
          "short_description": {
            "type": "object",
            "example": {
              "en": "Premium cotton polo",
              "vi": "Áo polo cotton cao cấp"
            },
            "description": "Short description in multiple languages"
          },
          "price": {
            "type": "number",
            "example": 399000
          },
          "sale_price": {
            "type": "number",
            "example": 349000
          },
          "cost_price": {
            "type": "number",
            "example": 200000
          },
          "images": {
            "example": [
              "https://example.com/image1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "stock_quantity": {
            "type": "number",
            "example": 50
          },
          "sku": {
            "type": "string",
            "example": "POLO-001"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890123"
          },
          "category_id": {
            "type": "string",
            "example": "b4b2b07f-6825-402b-bd2c-f9aef8cfbba5"
          },
          "tags": {
            "example": [
              "polo",
              "men",
              "premium"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "draft",
              "out_of_stock",
              "discontinued"
            ]
          },
          "is_featured": {
            "type": "boolean",
            "example": false
          },
          "enable_sale_tag": {
            "type": "boolean",
            "example": false
          },
          "meta_title": {
            "type": "object",
            "example": {
              "en": "Buy Premium Polo Shirt",
              "vi": "Mua Áo Polo Cao Cấp"
            },
            "description": "Meta title in multiple languages"
          },
          "meta_description": {
            "type": "object",
            "example": {
              "en": "High quality polo shirt...",
              "vi": "Áo polo chất lượng cao..."
            },
            "description": "Meta description in multiple languages"
          },
          "weight": {
            "type": "number",
            "example": 0.3,
            "description": "Weight in kg"
          },
          "dimensions": {
            "example": {
              "length": 28,
              "width": 20,
              "height": 2
            },
            "description": "Product dimensions in cm",
            "allOf": [
              {
                "$ref": "#/components/schemas/DimensionsDto"
              }
            ]
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "T-Shirts"
          },
          "slug": {
            "type": "string",
            "example": "t-shirts"
          },
          "description": {
            "type": "string",
            "example": "All types of t-shirts"
          },
          "image_url": {
            "type": "string",
            "example": "https://example.com/category.jpg"
          },
          "parent_id": {
            "type": "string",
            "example": "b4b2b07f-6825-402b-bd2c-f9aef8cfbba5"
          },
          "display_order": {
            "type": "number",
            "example": 1
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "T-Shirts"
          },
          "slug": {
            "type": "string",
            "example": "t-shirts"
          },
          "description": {
            "type": "string",
            "example": "All types of t-shirts"
          },
          "image_url": {
            "type": "string",
            "example": "https://example.com/category.jpg"
          },
          "parent_id": {
            "type": "string",
            "example": "b4b2b07f-6825-402b-bd2c-f9aef8cfbba5"
          },
          "display_order": {
            "type": "number",
            "example": 1
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}